refactor: group routing
