+++ src/api/tag/tag.ts
@@ -1,4 +1,4 @@
-import { Tag } from '@/app/[lng]/write/TagInput';
+import { Tag } from '@/app/[lng]/(common)/write/TagInput';
 import { gql } from '@/generated';
 
 import api from '../index';
