+++ src/app/[lng]/(common)/mypage/MypageSeperate.tsx
