+++ src/app/[lng]/(common)/mypage/MypageTable.tsx
