+++ src/app/[lng]/(common)/notice/[id]/Actions.tsx
