+++ src/app/[lng]/(common)/notice/[id]/AdditionalNotices.tsx
