+++ src/app/[lng]/(common)/notice/[id]/AuthorActions.tsx
