+++ src/app/[lng]/(common)/notice/[id]/Content.tsx
