+++ src/app/[lng]/(common)/notice/[id]/NoticeInfo.tsx
