+++ src/app/[lng]/(common)/notice/[id]/WriteEnglishNotice.tsx
