+++ src/app/[lng]/(common)/notice/[id]/loading.tsx
