+++ src/app/[lng]/(common)/notice/[id]/page.tsx
