+++ src/app/[lng]/(common)/page.tsx
@@ -1,7 +1,7 @@
 import { getAllNotices } from '@/api/notice/notice-server';
 
-import ZaboCarousel from '../components/templates/ZaboCarousel';
-import { createTranslation, PropsWithLng } from '../i18next';
+import ZaboCarousel from '../../components/templates/ZaboCarousel';
+import { createTranslation, PropsWithLng } from '../../i18next';
 
 export const dynamic = 'force-dynamic';
