+++ src/app/[lng]/(common)/section/[type]/layout.tsx
