+++ src/app/[lng]/(common)/section/[type]/page.tsx
