+++ src/app/[lng]/(common)/write/AttatchPhotoArea.tsx
