+++ src/app/[lng]/(common)/write/AttatchedPhoto.tsx
