+++ src/app/[lng]/(common)/write/handle-notice-submit.ts
