+++ src/app/[lng]/(common)/write/page.tsx
