+++ src/app/[lng]/(empty)/app/page.tsx
@@ -1,0 +1,8 @@
+import { headers } from 'next/headers';
+
+const AppOpenPage = () => {
+  const h = headers();
+  return <>{JSON.stringify(h)}</>;
+};
+
+export default AppOpenPage;
